+++ lib/src/layout.dart
@@ -26,17 +26,13 @@ class BottomNavLayout extends StatefulWidget {
     this.pageStack,
     this.extendBody = false,
     this.resizeToAvoidBottomInset = true,
-    this.pageTransitionBuilder,
-    this.pageTransitionDuration = 150,
-    this.pageTransitionDirection = AnimationDirection.inAndOut,
+    this.pageTransitionData,
   })  : assert(pages.length >= 1, "At least 1 page is required"),
-        //assert(pages.length == navBarDelegate.itemLength(), "Pass as many bottom navbar items as pages"), TODO: ?
-        assert(pageStack == null || pages.length > pageStack.peek() && pageStack.peek() >= 0, "initialPageIndex cannot exceed the page number or be negative"),
+        assert(
+            pageStack == null ||
+                pages.length > pageStack.peek() && pageStack.peek() >= 0,
+            "initialPageIndex cannot exceed the page number or be negative"),
         super(key: key);
-
-  final AnimatedBuilder Function(AnimationController, Widget)? pageTransitionBuilder;
-  final int pageTransitionDuration;
-  final AnimationDirection pageTransitionDirection;
 
   /// The app's destinations.
   /// Each destination corresponds to one bottom navbar item.
@@ -50,6 +46,9 @@ class BottomNavLayout extends StatefulWidget {
   final List<PageBuilder> pages;
 
   /// The bottom navbar of the layout.
+  ///
+  /// Make sure bottom navbar items count is the same as the page count.
+  /// Otherwise, you might get a [RangeError].
   final Widget Function(int, Function(int)) bottomNavigationBar;
 
   /// When false, the pages are reinitialized every time they are navigated to. (Material Design behavior)
@@ -74,6 +73,10 @@ class BottomNavLayout extends StatefulWidget {
 
   /// Passed to [Scaffold.resizeToAvoidBottomInset]. Default is true.
   final bool resizeToAvoidBottomInset;
+
+  /// Page transition data delegate.
+  /// If null, no transition is applied.
+  final PageTransitionData? pageTransitionData;
 
   @override
   State<StatefulWidget> createState() => _BottomNavLayoutState();
@@ -102,7 +105,10 @@ class _BottomNavLayoutState extends State<BottomNavLayout> {
   @override
   void initState() {
     // Set the pageStack. If not passed in, initialize with default.
-    pageStack = widget.pageStack ?? (Platform.isAndroid ? ReorderToFrontPageStack(initialPage: 0) : NoPageStack(initialPage: 0));
+    pageStack = widget.pageStack ??
+        (Platform.isAndroid
+            ? ReorderToFrontPageStack(initialPage: 0)
+            : NoPageStack(initialPage: 0));
 
     setState(() {
       // Initialize keys.
@@ -122,6 +128,10 @@ class _BottomNavLayoutState extends State<BottomNavLayout> {
   /// If the selected page is the current page, pops the page until it reaches it's root route.
   /// If the selected page is not the current page, navigates to that page.
   void onPageSelected(int index) {
+    if (index >= pages.length)
+      throw ArgumentError(
+          "There is no page corresponding to the selected navbar item at index $index.");
+
     // If the current item is selected
     if (index == pageStack.peek()) {
       // Pop until the base route
@@ -142,7 +152,8 @@ class _BottomNavLayoutState extends State<BottomNavLayout> {
   /// If there is a single page in the stack, bubbles up the pop event. Exits the app if no other back button handler is configured in the app.
   Future<bool> onWillPop() async {
     // Send pop event to the inner page
-    final consumedByPage = await keys[pageStack.peek()].currentState?.maybePop() ?? false;
+    final consumedByPage =
+        await keys[pageStack.peek()].currentState?.maybePop() ?? false;
 
     // If the back event is consumed by the inner page
     if (consumedByPage) {
@@ -183,25 +194,24 @@ class _BottomNavLayoutState extends State<BottomNavLayout> {
       child: Scaffold(
         extendBody: widget.extendBody,
         resizeToAvoidBottomInset: widget.resizeToAvoidBottomInset,
-        // Depending on if the user wants to save the page states
-        body: !widget.savePageState
-            // Do not save page states
-            ? pages[currentIndex]
-            : widget.pageTransitionBuilder == null
-                ? IndexedStack(
-                    index: currentIndex,
-                    // If the page is not initialized, "not show" an invisible widget instead.
-                    children: pages.map((page) => page ?? SizedBox.shrink()).toList(),
-                  )
-                : TwoWayAnimatedIndexedStack(
-                    animationBuilder: widget.pageTransitionBuilder!,
-                    animationDuration: widget.pageTransitionDuration,
-                    animationDirection: widget.pageTransitionDirection,
-                    index: currentIndex,
-                    // If the page is not initialized, "not show" an invisible widget instead.
-                    children: pages.map((page) => page ?? SizedBox.shrink()).toList(),
-                  ),
-        bottomNavigationBar: widget.bottomNavigationBar(currentIndex, onPageSelected),
+        body: widget.pageTransitionData == null
+            ? IndexedStack(
+                key: ValueKey<int>(widget.savePageState ? 0 : currentIndex),
+                index: currentIndex,
+                // If the page is not initialized, "not show" an invisible widget instead.
+                children:
+                    pages.map((page) => page ?? SizedBox.shrink()).toList(),
+              )
+            : TwoWayAnimatedIndexedStack(
+                key: ValueKey<int>(widget.savePageState ? 0 : currentIndex),
+                animData: widget.pageTransitionData!,
+                index: currentIndex,
+                // If the page is not initialized, "not show" an invisible widget instead.
+                children:
+                    pages.map((page) => page ?? SizedBox.shrink()).toList(),
+              ),
+        bottomNavigationBar:
+            widget.bottomNavigationBar(currentIndex, onPageSelected),
       ),
     );
   }