+++ CHANGELOG.md
@@ -4,6 +4,15 @@
    - `pages`, `pageBuilders` and `keys` parameters are combined into `pages` parameter. (breaking)
    - User don't need to create keys anymore.
    - Helper methods and documentation added for in-page navigation.
+ - Bottom bar delegates are removed.
+   - Bottom navbar is created by the user, using a builder.
+   - Removes one layer of API.
+   - Removes all bottom bar dependencies.
+ - Parameter defaults are changed to make the user more aware of their design choices.
+   - `savePageState` default is change to false.
+   - `pageStack` default is change to `NoPageStack` for iOS apps.
+ - State management inside library is improved.
+ - Documentation updated.
 
 ## 2.1.1
  - Jul 14, 2021